# [POAE7-2846]optimize compiler flags for release mode, support sccache