--- a/cpp/src/cider/exec/nextgen/context/CodegenContext.h
+++ b/cpp/src/cider/exec/nextgen/context/CodegenContext.h
@@ -71,8 +71,9 @@ struct CodegenOptions {
   bool needs_error_check = false;
   bool check_bit_vector_clear_opt = false;
   bool set_null_bit_vector_opt = false;
-  bool branchless_logic = false;
+  bool branchless_logic = true;
   bool enable_vectorize = false;
+
   jitlib::CompilationOptions co = jitlib::CompilationOptions{};
 };