--- a/cpp/src/cider/exec/nextgen/operators/ColumnToRowNode.cpp
+++ b/cpp/src/cider/exec/nextgen/operators/ColumnToRowNode.cpp
@@ -196,7 +196,7 @@ void ColumnToRowTranslator::codegenImpl(SuccessorEmitter successor_wrapper,
   auto len = func->createLocalJITValue([&input_array]() {
     return context::codegen_utils::getArrowArrayLength(input_array);
   });
-  if (!for_null) {
+  if (!for_null_) {
     static_cast<ColumnToRowNode*>(node_.get())->setColumnRowNum(len);
   }
   auto idx_upper = func->createVariable(JITTypeTag::INT64, "idx_upper", len);