### What changes were proposed in this pull request?
1. optimize compiler flags for release mode, which can generate fast `popcnt` asm code
    performance can boost about 10%~20%
2. support sccache, to accelerate compilation

performance compararion update:

```base
============================================================================
/workspace/code/BDTK/cpp/src/cider-velox/benchmark/expression/ArithmeticAndComparison.cpprelative  time/iter  iters/s
============================================================================
velox                                                         4.28s  233.70m
nextgen                                          104.06%      4.11s  243.20m
----------------------------------------------------------------------------
nextgen____base                                               4.04s  247.55m
nextgen_opt                                      321.11%      1.26s  794.91m
specifialize1                                   1550.02%   260.62ms     3.84
specifialize2                                   1070.15%   377.48ms     2.65
specifialize3                                    238.90%      1.69s  591.40m
----------------------------------------------------------------------------
mulI8Velox_________Base                                       3.88s  257.55m
mulI8NextGen                                      95.28%      4.08s  245.40m
mulI8NextGen_______Base                                       3.99s  250.75m
mulI8NextgenIsBitClear                            92.16%      4.33s  231.10m
mulI8NextgenBranchlessSetNull                     36.49%     10.93s   91.51m
mulI8NextgenNullSeparate                         319.79%      1.25s  801.88m
mulI8NextgenAllOpt                               325.60%      1.22s  816.43m
mulI8Compile                                                52.64ms    19.00
----------------------------------------------------------------------------
mulI8NestedVelox_________Base                                 6.63s  150.87m
mulI8NestedNextGen                               156.55%      4.23s  236.20m
mulI8NestedNextGen_______Base                                 4.31s  231.93m
mulI8NestedNextgenIsBitClear                      95.01%      4.54s  220.35m
mulI8NestedNextgenBranchlessSetNull               81.77%      5.27s  189.64m
mulI8NestedNextgenNullSeparate                   310.64%      1.39s  720.47m
mulI8NestedNextgenAllOpt                         291.66%      1.48s  676.43m
mulI8NestedCompile                                          52.56ms    19.03
----------------------------------------------------------------------------
mulI8NestedDeepVelox_________Base                             9.24s  108.24m
mulI8NestedDeepNextGen                           211.69%      4.36s  229.13m
mulI8NestedDeepNextGen_______Base                             4.37s  228.74m
mulI8NestedDeepNextgenIsBitClear                  96.50%      4.53s  220.72m
mulI8NestedDeepNextgenBranchlessSetNull           77.61%      5.63s  177.52m
mulI8NestedDeepNextgenNullSeparate               344.63%      1.27s  788.30m
mulI8NestedDeepNextgenAllOpt                     322.18%      1.36s  736.93m
mulI8NestedDeepCompile                                      52.63ms    19.00
----------------------------------------------------------------------------
mulI16Velox_________Base                                      3.40s  293.80m
mulI16NextGen                                     83.24%      4.09s  244.57m
mulI16NextGen_______Base                                      4.13s  241.99m
mulI16NextgenIsBitClear                           93.29%      4.43s  225.76m
mulI16NextgenBranchlessSetNull                    38.34%     10.78s   92.78m
mulI16NextgenNullSeparate                        315.88%      1.31s  764.42m
mulI16NextgenAllOpt                              323.82%      1.28s  783.63m
mulI16Compile                                               52.83ms    18.93
----------------------------------------------------------------------------
mulI32Velox_________Base                                      3.35s  298.60m
mulI32NextGen                                     80.72%      4.15s  241.03m
mulI32NextGen_______Base                                      4.18s  239.13m
mulI32NextgenIsBitClear                           93.85%      4.46s  224.42m
mulI32NextgenBranchlessSetNull                    38.29%     10.92s   91.57m
mulI32NextgenNullSeparate                        281.62%      1.48s  673.44m
mulI32NextgenAllOpt                              265.73%      1.57s  635.46m
mulI32Compile                                               53.54ms    18.68
----------------------------------------------------------------------------
mulI64Velox_________Base                                      3.51s  284.52m
mulI64NextGen                                     82.04%      4.28s  233.43m
mulI64NextGen_______Base                                      4.27s  234.12m
mulI64NextgenIsBitClear                           91.58%      4.66s  214.41m
mulI64NextgenBranchlessSetNull                    38.82%     11.00s   90.87m
mulI64NextgenNullSeparate                        178.03%      2.40s  416.81m
mulI64NextgenAllOpt                              190.78%      2.24s  446.65m
mulI64Compile                                               53.13ms    18.82
----------------------------------------------------------------------------
mulDoubleVelox_________Base                                   3.36s  297.31m
mulDoubleNextGen                                  69.79%      4.82s  207.51m
mulDoubleNextGen_______Base                                   4.95s  202.15m
mulDoubleNextgenIsBitClear                        85.73%      5.77s  173.30m
mulDoubleNextgenBranchlessSetNull                 56.50%      8.75s  114.22m
mulDoubleNextgenNullSeparate                     237.52%      2.08s  480.14m
mulDoubleNextgenAllOpt                           234.52%      2.11s  474.09m
mulDoubleCompile                                            53.60ms    18.66
----------------------------------------------------------------------------
mulDoubleSameColumnVelox_________Base                         3.42s  292.59m
mulDoubleSameColumnNextGen                        76.63%      4.46s  224.20m
mulDoubleSameColumnNextGen_______Base                         4.28s  233.66m
mulDoubleSameColumnNextgenIsBitClear              93.28%      4.59s  217.96m
mulDoubleSameColumnNextgenBranchlessSetNull       80.27%      5.33s  187.56m
mulDoubleSameColumnNextgenNullSeparate           259.02%      1.65s  605.23m
mulDoubleSameColumnNextgenAllOpt                 224.24%      1.91s  523.97m
mulDoubleSameColumnCompile                                  52.89ms    18.91
----------------------------------------------------------------------------
mulDoubleConstantVelox_________Base                           3.31s  302.02m
mulDoubleConstantNextGen                          67.32%      4.92s  203.32m
mulDoubleConstantNextGen_______Base                           4.94s  202.33m
mulDoubleConstantNextgenIsBitClear                86.06%      5.74s  174.12m
mulDoubleConstantNextgenBranchlessSetNull         56.55%      8.74s  114.42m
mulDoubleConstantNextgenNullSeparate             243.88%      2.03s  493.45m
mulDoubleConstantNextgenAllOpt                   237.93%      2.08s  481.40m
mulDoubleConstantCompile                                    53.63ms    18.65
----------------------------------------------------------------------------
mulDoubleNestedVelox_________Base                             4.94s  202.42m
mulDoubleNestedNextGen                            97.96%      5.04s  198.29m
mulDoubleNestedNextGen_______Base                             5.07s  197.42m
mulDoubleNestedNextgenIsBitClear                  85.35%      5.93s  168.50m
mulDoubleNestedNextgenBranchlessSetNull           79.06%      6.41s  156.07m
mulDoubleNestedNextgenNullSeparate               245.22%      2.07s  484.11m
mulDoubleNestedNextgenAllOpt                     235.40%      2.15s  464.73m
mulDoubleNestedCompile                                      53.97ms    18.53
----------------------------------------------------------------------------
mulDoubleNestedDeepVelox_________Base                         9.79s  102.13m
mulDoubleNestedDeepNextGen                       200.61%      4.88s  204.89m
mulDoubleNestedDeepNextGen_______Base                         5.00s  199.87m
mulDoubleNestedDeepNextgenIsBitClear              82.92%      6.03s  165.74m
mulDoubleNestedDeepNextgenBranchlessSetNull       74.59%      6.71s  149.08m
mulDoubleNestedDeepNextgenNullSeparate           226.57%      2.21s  452.86m
mulDoubleNestedDeepNextgenAllOpt                 234.69%      2.13s  469.07m
mulDoubleNestedDeepCompile                                  54.19ms    18.46
----------------------------------------------------------------------------
PlusCheckedVeloxI64                                           4.46s  224.06m
plusI64Velox_________Base                                     3.57s  279.94m
plusI64NextGen                                    83.11%      4.30s  232.64m
plusI64NextGen_______Base                                     4.37s  229.09m
plusI64NextgenIsBitClear                          95.51%      4.57s  218.79m
plusI64NextgenBranchlessSetNull                   39.72%     10.99s   90.99m
plusI64NextgenNullSeparate                       294.61%      1.48s  674.94m
plusI64NextgenAllOpt                             249.42%      1.75s  571.39m
plusI64Compile                                              53.26ms    18.78
----------------------------------------------------------------------------
mulAndAddVelox_________Base                                  16.16s   61.89m
mulAndAddNextGen                                 312.43%      5.17s  193.35m
mulAndAddNextGen_______Base                                   5.33s  187.63m
mulAndAddNextgenIsBitClear                       100.96%      5.28s  189.44m
mulAndAddNextgenBranchlessSetNull                 84.43%      6.31s  158.42m
mulAndAddNextgenNullSeparate                     242.29%      2.20s  454.60m
mulAndAddNextgenAllOpt                           235.63%      2.26s  442.11m
mulAndAddCompile                                            54.07ms    18.50
----------------------------------------------------------------------------
eqVelox_________Base                                          3.86s  259.19m
eqNextGen                                         67.56%      5.71s  175.12m
eqNextGen_______Base                                          5.72s  174.70m
eqNextgenIsBitClear                               90.14%      6.35s  157.48m
eqNextgenBranchlessSetNull                        51.66%     11.08s   90.24m
eqNextgenNullSeparate                            171.36%      3.34s  299.38m
eqNextgenAllOpt                                  121.65%      4.71s  212.53m
eqCompile                                                   53.93ms    18.54
----------------------------------------------------------------------------
eqConstantVelox_________Base                                  3.90s  256.36m
eqConstantNextGen                                 68.18%      5.72s  174.80m
eqConstantNextGen_______Base                                  5.80s  172.52m
eqConstantNextgenIsBitClear                       94.21%      6.15s  162.52m
eqConstantNextgenBranchlessSetNull                51.98%     11.15s   89.68m
eqConstantNextgenNullSeparate                    177.47%      3.27s  306.16m
eqConstantNextgenAllOpt                          120.71%      4.80s  208.25m
eqConstantCompile                                           53.72ms    18.61
----------------------------------------------------------------------------
eqBoolVelox_________Base                                      4.39s  227.60m
eqBoolNextGen                                     73.49%      5.98s  167.27m
eqBoolNextGen_______Base                                      5.95s  167.99m
eqBoolNextgenIsBitClear                           78.67%      7.57s  132.17m
eqBoolNextgenBranchlessSetNull                    65.15%      9.14s  109.45m
eqBoolNextgenNullSeparate                        135.04%      4.41s  226.85m
eqBoolNextgenAllOpt                              108.13%      5.50s  181.65m
eqBoolCompile                                               54.55ms    18.33
----------------------------------------------------------------------------
neqVelox_________Base                                         3.67s  272.54m
neqNextGen                                        69.97%      5.24s  190.69m
neqNextGen_______Base                                         5.29s  189.13m
neqNextgenIsBitClear                              83.55%      6.33s  158.03m
neqNextgenBranchlessSetNull                       47.22%     11.20s   89.30m
neqNextgenNullSeparate                           158.97%      3.33s  300.67m
neqNextgenAllOpt                                 116.36%      4.54s  220.08m
neqCompile                                                  53.71ms    18.62
----------------------------------------------------------------------------
gtVelox_________Base                                          3.76s  266.29m
gtNextGen                                         68.84%      5.45s  183.32m
gtNextGen_______Base                                          5.48s  182.59m
gtNextgenIsBitClear                               81.21%      6.74s  148.27m
gtNextgenBranchlessSetNull                        69.82%      7.84s  127.48m
gtNextgenNullSeparate                            163.83%      3.34s  299.13m
gtNextgenAllOpt                                  114.12%      4.80s  208.36m
gtCompile                                                   53.65ms    18.64
----------------------------------------------------------------------------
ltVelox_________Base                                          3.63s  275.32m
ltNextGen                                         66.72%      5.44s  183.70m
ltNextGen_______Base                                          5.47s  182.87m
ltNextgenIsBitClear                               81.41%      6.72s  148.86m
ltNextgenBranchlessSetNull                        68.65%      7.97s  125.54m
ltNextgenNullSeparate                            160.45%      3.41s  293.41m
ltNextgenAllOpt                                  115.72%      4.73s  211.61m
ltCompile                                                   53.79ms    18.59
----------------------------------------------------------------------------
andVelox_________Base                                        40.34s   24.79m
andNextGen                                       500.18%      8.06s  124.00m
andNextGen_______Base                                         7.92s  126.29m
andNextgenIsBitClear                              88.46%      8.95s  111.72m
andNextgenBranchlessSetNull                       81.86%      9.67s  103.38m
andNextgenNullSeparate                           208.30%      3.80s  263.07m
andNextgenAllOpt                                 146.18%      5.42s  184.62m
andCompile                                                  54.78ms    18.25
----------------------------------------------------------------------------
orVelox_________Base                                         48.47s   20.63m
orNextGen                                        709.05%      6.84s  146.27m
orNextGen_______Base                                          6.92s  144.48m
orNextgenIsBitClear                               79.65%      8.69s  115.07m
orNextgenBranchlessSetNull                        76.29%      9.07s  110.22m
orNextgenNullSeparate                            167.09%      4.14s  241.40m
orNextgenAllOpt                                   83.70%      8.27s  120.93m
orCompile                                                   54.82ms    18.24
----------------------------------------------------------------------------
conjunctsNestedVelox_________Base                           3.17min    5.25m
conjunctsNestedNextGen                          1120.08%     17.00s   58.83m
conjunctsNestedNextGen_______Base                            17.20s   58.14m
conjunctsNestedNextgenIsBitClear                 112.56%     15.28s   65.45m
conjunctsNestedNextgenBranchlessSetNull           67.76%     25.38s   39.40m
conjunctsNestedNextgenNullSeparate               216.37%      7.95s  125.80m
conjunctsNestedNextgenAllOpt                     228.28%      7.53s  132.73m
conjunctsNestedCompile                                      57.18ms    17.49
----------------------------------------------------------------------------
```

### Why are the changes needed?
performance for run and compile.


### Does this PR introduce _any_ user-facing change?
No


### How was this patch tested?
No need

### Which label does this PR belong to?
INFRA